requirements.txt
pip
Django
django-annoying